Adding an in-place editor session to the spell checker script
Now, when users of the spell checker are confronted with a misspelt word, they are given the option of opening the file in vim to edit it then and there (like a git commit message).  The file opens and the misspelt word is already highlighted, everywhere that it appears.

We need to re-execute the spell checker after the editor session
terminates... in the next commit, perhaps!

In addition, it might be nice to give a straightforward "replace" option, which uses some unix tools to regex-replace the word, everywhere that it appears in the file.